# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Post Your Opinion API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get list of users",
                "description": "Returns list of users",
                "operationId": "get all users ",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "users": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "user id",
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "description": "user name",
                                                        "type": "string"
                                                    },
                                                    "email": {
                                                        "description": "user email",
                                                        "type": "string"
                                                    },
                                                    "email_verified_at": {
                                                        "description": "user email verified timestamp",
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "description": "user creation timestamp",
                                                        "type": "string"
                                                    },
                                                    "updated_at": {
                                                        "description": "user updated timestamp",
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "collectionFormat": "multi"
                                        },
                                        "message": {
                                            "description": "Message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create new user",
                "description": "Create new user",
                "operationId": "create user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "user name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "user email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "user password",
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "description": "user password confirmation",
                                        "type": "string"
                                    },
                                    "role": {
                                        "description": "user role",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "user id",
                                            "type": "integer"
                                        },
                                        "name": {
                                            "description": "user name",
                                            "type": "string"
                                        },
                                        "email": {
                                            "description": "user email",
                                            "type": "string"
                                        },
                                        "email_verified_at": {
                                            "description": "user email verified timestamp",
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "description": "user creation timestamp",
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "description": "user updated timestamp",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}